Simplify Hook Initializer
