+++ app/src/main/java/org/kitsunepie/qassistant/app/HookLoader.kt
@@ -1,7 +1,7 @@
 package org.kitsunepie.qassistant.app
 
 import com.github.kyuubiran.ezxhelper.utils.Log
-import org.kitsunepie.qassistant.app.hook.base.BaseModuleInitHook
+import org.kitsunepie.qassistant.app.hook.base.HookInitializer
 
 object HookLoader {
     @JvmStatic
@@ -12,7 +12,7 @@ object HookLoader {
     private fun doInit() {
         try {
             //加载模块初始化hooks
-            BaseModuleInitHook.initHooks()
+            HookInitializer.initModuleHooks()
         } catch (e: Exception) {
             Log.e(e)
         }
